PHOENIX-5230 Fix ChangePermissionsIT and TableDDLPermissionIT on master
